+++ src/MEM.Zone-Dashboards/docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for MEM.Zone-Dashboards
 
+## 5.0.1 - 2021-10-14
+
+* Included `ReportingServicesTools` dependency in the package for the `Install-SRSReport` script.
+* Minor Documentation additions.
+
 ## 5.0.0 - 2021-10-13
 
 * Added `Operating System - Version Compliance by Collection` report.